pyproject.toml
[tool.poetry]
name = "metricone"
version = "0.1.0"
description = "Metric management system backend scaffold"
authors = ["MetricOne Team <dev@metricone.local>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
sqlalchemy = "^2.0.28"
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
psycopg = { extras = ["binary"], version = "^3.1.18" }
redis = "^5.0.3"
celery = "^5.3.6"
aiosqlite = "^0.20.0"
python-multipart = "^0.0.9"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.8.0"
minio = "^7.2.7"
loguru = "^0.7.2"
prometheus-client = "^0.20.0"
pandas = "^2.2.1"
pyarrow = "^15.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
ruff = "^0.3.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core>=1.8.2"]
build-backend = "poetry.core.masonry.api"
